% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_icpms.R
\name{read_icpms}
\alias{read_icpms}
\alias{tidy_icpms}
\title{Read ICP-MS Data Files}
\usage{
read_icpms(path, skip = 1, na = c("NA", "", "n/a"))

tidy_icpms(df)
}
\arguments{
\item{path}{The path where the files are kept}

\item{skip}{The number of lines to skip at the beginning of the file
(blank lines are automatically dropped).}

\item{na}{Values that should be considered NA}

\item{df}{The output of \code{read_icpms} or a filtered result}
}
\value{
\code{read_icpms} returns a
  data frame with columns sample_id, datetime, run, and others in
  the form "27Al_ppb" (isotope, element, unit). \code{tidy_icpms} returns
  a parameter-long, summarised result.
}
\description{
Read ICP-MS Data Files
}
\examples{
# file
read_icpms(system.file("test_icpms/test_icpms.xlsx", package = "dalcwrs"))
# directory
read_icpms(system.file("test_icpms", package = "dalcwrs"))

# use tidy to get parameter-long, summarised output
raw <- read_icpms(system.file("test_icpms/test_icpms.xlsx", package = "dalcwrs"))
tidy_icpms(raw)

}
